Add strategy class
