+++ app/src/main/java/io/github/wulkanowy/data/repositories/preferences/PreferencesRepository.kt
@@ -18,7 +18,7 @@ class PreferencesRepository @Inject constructor(
         get() = sharedPref.getBoolean(context.getString(R.string.pref_key_attendance_present), true)
 
     val gradeAverageMode: String
-        get() = sharedPref.getString(context.getString(R.string.pref_key_grade_average_mode), "only_last_semester") ?: "only_last_semester"
+        get() = sharedPref.getString(context.getString(R.string.pref_key_grade_average_mode), "only_one_semester") ?: "only_one_semester"
 
     val isGradeExpandable: Boolean
         get() = !sharedPref.getBoolean(context.getString(R.string.pref_key_expand_grade), false)
